Move `Enumerable` functionality to `EnumerableObject`
