+++ lib/redis/helpers/core_commands.rb
@@ -50,12 +50,6 @@ class Redis
         redis.move key, dbindex
       end
 
-      def sort(options={})
-        options[:order] = "asc alpha" if options.keys.count == 0  # compat with Ruby
-        val = redis.sort(key, options)
-        val.is_a?(Array) ? val.map{|v| unmarshal(v)} : val
-      end
-
       def marshal(value, domarshal=false)
         if options[:marshal] || domarshal
           Marshal.dump(value)
